Move remaining embedded study tests to functional tests
Again, I don't think these are quite the tests we want but I'd rather
move them as they are and think about changing them later.
