--- tests/acceptance/test_embedded_study.py
@@ -1,202 +1,0 @@
-from datetime import datetime
-
-import pytest
-
-from ehrql.file_formats import FILE_FORMATS
-from tests.lib.inspect_utils import function_body_as_string
-from tests.lib.tpp_schema import AllowedPatientsWithTypeOneDissent, Patient
-
-
-@function_body_as_string
-def trivial_dataset_definition():
-    from ehrql import create_dataset
-    from ehrql.tables.tpp import patients
-
-    dataset = create_dataset()
-    year = patients.date_of_birth.year
-    dataset.define_population(year >= 1940)
-    dataset.year = year
-
-    dataset.configure_dummy_data(
-        population_size=10,
-        additional_population_constraint=patients.date_of_death.is_null(),
-    )
-
-
-@function_body_as_string
-def trivial_dataset_definition_legacy_dummy_data():
-    from ehrql import create_dataset
-    from ehrql.tables.tpp import patients
-
-    dataset = create_dataset()
-    year = patients.date_of_birth.year
-    dataset.define_population(year >= 1940)
-    dataset.year = year
-
-    dataset.configure_dummy_data(population_size=10, legacy=True)
-
-
-@function_body_as_string
-def parameterised_dataset_definition():
-    from argparse import ArgumentParser
-
-    from ehrql import create_dataset
-    from ehrql.tables.tpp import patients
-
-    parser = ArgumentParser()
-    parser.add_argument("--year", type=int)
-    args = parser.parse_args()
-
-    dataset = create_dataset()
-    year = patients.date_of_birth.year
-    dataset.define_population(year >= args.year)
-    dataset.year = year
-
-
-@pytest.mark.parametrize("extension", list(FILE_FORMATS.keys()))
-def test_generate_dataset(study, mssql_database, extension):
-    mssql_database.setup(
-        Patient(Patient_ID=1, DateOfBirth=datetime(1934, 5, 5)),
-        AllowedPatientsWithTypeOneDissent(Patient_ID=1),
-        Patient(Patient_ID=2, DateOfBirth=datetime(1943, 5, 5)),
-        AllowedPatientsWithTypeOneDissent(Patient_ID=2),
-        Patient(Patient_ID=3, DateOfBirth=datetime(1999, 5, 5)),
-        AllowedPatientsWithTypeOneDissent(Patient_ID=3),
-    )
-
-    study.setup_from_string(trivial_dataset_definition)
-    study.generate(mssql_database, "ehrql.backends.tpp.TPPBackend", extension=extension)
-    results = study.results()
-
-    expected = [1943, 1999]
-    if extension in (".csv", ".csv.gz"):
-        expected = [str(v) for v in expected]
-
-    assert len(results) == len(expected)
-    assert {r["year"] for r in results} == set(expected)
-
-
-def test_parameterised_dataset_definition(study, mssql_database):
-    mssql_database.setup(
-        Patient(Patient_ID=1, DateOfBirth=datetime(1934, 5, 5)),
-        AllowedPatientsWithTypeOneDissent(Patient_ID=1),
-        Patient(Patient_ID=2, DateOfBirth=datetime(1943, 5, 5)),
-        AllowedPatientsWithTypeOneDissent(Patient_ID=2),
-        Patient(Patient_ID=3, DateOfBirth=datetime(1999, 5, 5)),
-        AllowedPatientsWithTypeOneDissent(Patient_ID=3),
-    )
-
-    study.setup_from_string(parameterised_dataset_definition)
-    study.generate(
-        mssql_database,
-        "ehrql.backends.tpp.TPPBackend",
-        user_args={"year": "1940"},
-    )
-    results = study.results()
-
-    expected = ["1943", "1999"]
-
-    assert len(results) == len(expected)
-    assert {r["year"] for r in results} == set(expected)
-
-
-def test_parameterised_dataset_definition_with_bad_param(study, mssql_database, capsys):
-    study.setup_from_string(parameterised_dataset_definition)
-    with pytest.raises(SystemExit):
-        study.generate(
-            mssql_database,
-            "ehrql.backends.tpp.TPPBackend",
-            user_args={"ear": "1940"},
-        )
-    assert (
-        "dataset.py: error: unrecognized arguments: --ear 1940"
-        in capsys.readouterr().err
-    )
-
-
-def test_generate_dataset_with_database_error(study, mssql_database):
-    mssql_database.setup(
-        Patient(Patient_ID=1, DateOfBirth=datetime(1934, 5, 5)),
-        AllowedPatientsWithTypeOneDissent(Patient_ID=1),
-    )
-
-    @function_body_as_string
-    def database_operational_error_dataset_definition():
-        from ehrql import create_dataset, years
-        from ehrql.tables.core import patients
-
-        dataset = create_dataset()
-        dataset.define_population(patients.date_of_birth.year >= 1900)
-        dataset.extended_dob = patients.date_of_birth + years(9999)
-
-        dataset.configure_dummy_data(population_size=10)
-
-    # This dataset definition triggers an OperationalError by implementing date
-    # arithmetic that results in an out of bounds date (after 9999-12-31)
-    study.setup_from_string(database_operational_error_dataset_definition)
-    with pytest.raises(SystemExit) as err:
-        study.generate(mssql_database, "ehrql.backends.tpp.TPPBackend")
-    assert err.value.code == 5
-
-
-def test_validate_dummy_data_happy_path(study, tmp_path):
-    dummy_data_file = tmp_path / "dummy.csv"
-    dummy_data = "patient_id,year\n1,1971\n2,1992"
-    dummy_data_file.write_text(dummy_data)
-    study.setup_from_string(trivial_dataset_definition)
-    study.generate(
-        database=None,
-        backend="expectations",
-        dummy_data_file=str(dummy_data_file),
-    )
-    assert study.results() == [
-        {"patient_id": "1", "year": "1971"},
-        {"patient_id": "2", "year": "1992"},
-    ]
-
-
-def test_validate_dummy_data_error_path(study, tmp_path, capsys):
-    dummy_data_file = tmp_path / "dummy.csv"
-    dummy_data = "patient_id,year\n1,1971\n2,foo"
-    dummy_data_file.write_text(dummy_data)
-    study.setup_from_string(trivial_dataset_definition)
-    with pytest.raises(SystemExit):
-        study.generate(
-            database=None,
-            backend="expectations",
-            dummy_data_file=str(dummy_data_file),
-        )
-    captured = capsys.readouterr()
-    assert "invalid literal for int" in captured.err
-
-
-@pytest.mark.parametrize(
-    "dataset_definition_fixture",
-    (
-        trivial_dataset_definition,
-        trivial_dataset_definition_legacy_dummy_data,
-    ),
-)
-def test_generate_dummy_data(study, dataset_definition_fixture):
-    study.setup_from_string(dataset_definition_fixture)
-    study.generate(database=None, backend="expectations", extension=".csv")
-    lines = study._dataset_path.read_text().splitlines()
-    assert lines[0] == "patient_id,year"
-    assert len(lines) == 11  # 1 header, 10 rows
-
-
-def test_generate_dummy_data_with_dummy_tables(study, tmp_path):
-    tmp_path.joinpath("patients.csv").write_text(
-        "patient_id,date_of_birth\n8,1985-10-20\n9,1995-05-10"
-    )
-    study.setup_from_string(trivial_dataset_definition)
-    study.generate(
-        database=None,
-        backend="expectations",
-        extension=".csv",
-        dummy_tables=str(tmp_path),
-    )
-    assert study.results() == [
-        {"patient_id": "8", "year": "1985"},
-        {"patient_id": "9", "year": "1995"},
-    ]
